Move runtime-specific DLL injection code into
