+++ dnSpy.Extension.HoLLy/CodeInjection/Injectors/IInjector.cs
@@ -1,0 +1,7 @@
+namespace HoLLy.dnSpyExtension.CodeInjection.Injectors
+{
+    public interface IInjector
+    {
+        void Inject(int pid, string path, string typeName, string methodName, string? parameter, bool x86);
+    }
+}
